+++ client/src/core/components/App.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { Route, BrowserRouter as Router } from 'react-router-dom';
-import Dashboard from './dashboard/Dashboard';
-import Submission from './submission/Submission';
-import NavBar from './core/NavBar';
+import Dashboard from '../../dashboard/components/Dashboard';
+import Submission from '../../submission/Submission';
+import NavBar from './NavBar';
 
 import '../styles/index.scss'
