+++ client/src/core/styles/index.scss
@@ -2,5 +2,5 @@
 @import './_mixins.scss';
 
 //ui
-@import './atoms/index';
+@import '../../ui/styles/index';
