+++ client/src/ui/styles/AppBar.scss
