+++ client/src/ui/styles/AppBarIcon.scss
